chore: move AddToCalendar, JoinSession, SesionAttrs & Session header to "core" module
Also fix various minor issue too
