+++ src/core/session/JoinSession.vue
