+++ src/core/session/SessionHeader.vue
