+++ src/schedule/module.ts
@@ -1,22 +1,14 @@
-// actions/
-export { default as AddToCalendar } from './actions/AddToCalendar.vue';
-export { default as JoinSession } from './actions/JoinSession.vue';
-
 // filtering/
 export type { FilterOption } from './filtering/filter-option';
 export { default as InlineFilter } from './filtering/InlineFilter.vue';
 export { default as ScheduleFilters } from './filtering/ScheduleFilters.vue';
-
-// tiles/
-export { default as SessionAttributes } from './tile/SessionAttributes.vue';
-export { default as SessionHeader } from './tile/SessionHeader.vue';
-export { default as SessionTile } from './tile/SessionTile.vue';
 
 export { default as NoResults } from './NoResults.vue';
 export { default as ScheduleBlock } from './ScheduleBlock.vue';
 export { default as ScheduleHeader } from './ScheduleHeader.vue';
 export { default as ScheduleView } from './ScheduleView.vue';
 export { default as SessionBoard } from './SessionBoard.vue';
+export { default as SessionTile } from './SessionTile.vue';
 export { default as SpeakerGrid } from './SpeakerGrid.vue';
 export { default as StickyHeading } from './StickyHeading.vue';
 export { default as TimeSlot } from './TimeSlot.vue';
