+++ src/session/SessionLayout.vue
@@ -34,7 +34,6 @@
 // - $sessionLayout-background
 //
 
-/** @deprecated This wasn't used, a combination of `UtilLayout` and `ApiContent` work well */
 export default {
   name: 'SessionLayout'
 };
